Working Network Class
Established a working model for the Networking class and created a
Tester class to test it. So far, the program can connect to a url to a
server and read the output. The url of the server will probably have to
be changed, when we get the server up and running at school